refactor: Car 클래스의 책임 분리(Name 클래스 생성)
